btrfs: declare max_inline as u32
As of now btrfs_fs_info::max_line is u64, which can't be larger than
btrfs_fs_info::sectorsize which is defined as u32, so make
btrfs_fs_info::max_line u32.

+++ fs/btrfs/ctree.h
@@ -796,7 +796,7 @@ struct btrfs_fs_info {
 	 * extent. The write side(mount/remount) is under ->s_umount lock,
 	 * so it is also safe.
 	 */
-	u64 max_inline;
+	u32 max_inline;
 
 	struct btrfs_transaction *running_transaction;
 	wait_queue_head_t transaction_throttle;
